IPL Visualization assignment - Tableau
The Sports Editor of the agency has approached you to build a Tableau dashboard of IPL statistics over the years since its inception in order to create an infographic for a newsletter that their team is working on. For this newsletter, in some cases, they will use the visual representations as you have created in Tableau directly for their infographic, and in a few other cases, they will use important statistics after trying out the different filters and customisations that you have provided for interactivity.